@@ -1,5 +1,6 @@
 import re
 from contextlib import contextmanager
+from typing import Tuple
 
 from parso.python.errors import ErrorFinder, ErrorFinderConfig
 from parso.normalizer import Rule
@@ -15,16 +16,17 @@ _CLOSING_BRACKETS = ')', ']', '}'
 _FACTOR = '+', '-', '~'
 _ALLOW_SPACE = '*', '+', '-', '**', '/', '//', '@'
 _BITWISE_OPERATOR = '<<', '>>', '|', '&', '^'
-_NEEDS_SPACE = ('=', '%', '->',
-                '<', '>', '==', '>=', '<=', '<>', '!=',
-                '+=', '-=', '*=', '@=', '/=', '%=', '&=', '|=', '^=', '<<=',
-                '>>=', '**=', '//=')
+_NEEDS_SPACE: Tuple[str, ...] = (
+    '=', '%', '->',
+    '<', '>', '==', '>=', '<=', '<>', '!=',
+    '+=', '-=', '*=', '@=', '/=', '%=', '&=', '|=', '^=', '<<=',
+    '>>=', '**=', '//=')
 _NEEDS_SPACE += _BITWISE_OPERATOR
 _IMPLICIT_INDENTATION_TYPES = ('dictorsetmaker', 'argument')
 _POSSIBLE_SLICE_PARENTS = ('subscript', 'subscriptlist', 'sliceop')
 
 
-class IndentationTypes(object):
+class IndentationTypes:
     VERTICAL_BRACKET = object()
     HANGING_BRACKET = object()
     BACKSLASH = object()
@@ -71,7 +73,6 @@ class BracketNode(IndentationNode):
             n = n.parent
         parent_indentation = n.indentation
 
-
         next_leaf = leaf.get_next_leaf()
         if '\n' in next_leaf.prefix:
             # This implies code like:
@@ -93,7 +94,7 @@ class BracketNode(IndentationNode):
             if '\t' in config.indentation:
                 self.indentation = None
             else:
-                self.indentation =  ' ' * expected_end_indent
+                self.indentation = ' ' * expected_end_indent
             self.bracket_indentation = self.indentation
             self.type = IndentationTypes.VERTICAL_BRACKET
 
@@ -111,7 +112,7 @@ class ImplicitNode(BracketNode):
     annotations and dict values.
     """
     def __init__(self, config, leaf, parent):
-        super(ImplicitNode, self).__init__(config, leaf, parent)
+        super().__init__(config, leaf, parent)
         self.type = IndentationTypes.IMPLICIT
 
         next_leaf = leaf.get_next_leaf()
@@ -137,7 +138,7 @@ class BackslashNode(IndentationNode):
                     self.indentation = parent_indentation + config.indentation
                 else:
                     # +1 because there is a space.
-                    self.indentation =  ' ' * (equals.end_pos[1] + 1)
+                    self.indentation = ' ' * (equals.end_pos[1] + 1)
         else:
             self.indentation = parent_indentation + config.indentation
         self.bracket_indentation = self.indentation
@@ -150,7 +151,7 @@ def _is_magic_name(name):
 
 class PEP8Normalizer(ErrorFinder):
     def __init__(self, *args, **kwargs):
-        super(PEP8Normalizer, self).__init__(*args, **kwargs)
+        super().__init__(*args, **kwargs)
         self._previous_part = None
         self._previous_leaf = None
         self._on_newline = True
@@ -173,7 +174,7 @@ class PEP8Normalizer(ErrorFinder):
 
     @contextmanager
     def visit_node(self, node):
-        with super(PEP8Normalizer, self).visit_node(node):
+        with super().visit_node(node):
             with self._visit_node(node):
                 yield
 
@@ -190,7 +191,8 @@ class PEP8Normalizer(ErrorFinder):
             expr_stmt = node.parent
             # Check if it's simply defining a single name, not something like
             # foo.bar or x[1], where using a lambda could make more sense.
-            if expr_stmt.type == 'expr_stmt' and any(n.type == 'name' for n in expr_stmt.children[:-2:2]):
+            if expr_stmt.type == 'expr_stmt' and any(n.type == 'name'
+                                                     for n in expr_stmt.children[:-2:2]):
                 self.add_issue(node, 731, 'Do not assign a lambda expression, use a def')
         elif typ == 'try_stmt':
             for child in node.children:
@@ -221,7 +223,6 @@ class PEP8Normalizer(ErrorFinder):
         if typ in _IMPORT_TYPES:
             simple_stmt = node.parent
             module = simple_stmt.parent
-            #if module.type == 'simple_stmt':
             if module.type == 'file_input':
                 index = module.children.index(simple_stmt)
                 for child in module.children[:index]:
@@ -341,7 +342,7 @@ class PEP8Normalizer(ErrorFinder):
         self._newline_count = 0
 
     def visit_leaf(self, leaf):
-        super(PEP8Normalizer, self).visit_leaf(leaf)
+        super().visit_leaf(leaf)
         for part in leaf._split_prefix():
             if part.type == 'spacing':
                 # This part is used for the part call after for.
@@ -406,7 +407,6 @@ class PEP8Normalizer(ErrorFinder):
                     and leaf.parent.parent.type == 'decorated':
                 self.add_issue(part, 304, "Blank lines found after function decorator")
 
-
             self._newline_count += 1
 
         if type_ == 'backslash':
@@ -461,33 +461,62 @@ class PEP8Normalizer(ErrorFinder):
                     else:
                         should_be_indentation = node.indentation
                     if self._in_suite_introducer and indentation == \
-                                node.get_latest_suite_node().indentation \
-                                + self._config.indentation:
-                            self.add_issue(part, 129, "Line with same indent as next logical block")
+                            node.get_latest_suite_node().indentation \
+                            + self._config.indentation:
+                        self.add_issue(part, 129, "Line with same indent as next logical block")
                     elif indentation != should_be_indentation:
                         if not self._check_tabs_spaces(spacing) and part.value != '\n':
                             if value in '])}':
                                 if node.type == IndentationTypes.VERTICAL_BRACKET:
-                                    self.add_issue(part, 124, "Closing bracket does not match visual indentation")
+                                    self.add_issue(
+                                        part,
+                                        124,
+                                        "Closing bracket does not match visual indentation"
+                                    )
                                 else:
-                                    self.add_issue(part, 123, "Losing bracket does not match indentation of opening bracket's line")
+                                    self.add_issue(
+                                        part,
+                                        123,
+                                        "Losing bracket does not match "
+                                        "indentation of opening bracket's line"
+                                    )
                             else:
                                 if len(indentation) < len(should_be_indentation):
                                     if node.type == IndentationTypes.VERTICAL_BRACKET:
-                                        self.add_issue(part, 128, 'Continuation line under-indented for visual indent')
+                                        self.add_issue(
+                                            part,
+                                            128,
+                                            'Continuation line under-indented for visual indent'
+                                        )
                                     elif node.type == IndentationTypes.BACKSLASH:
-                                        self.add_issue(part, 122, 'Continuation line missing indentation or outdented')
+                                        self.add_issue(
+                                            part,
+                                            122,
+                                            'Continuation line missing indentation or outdented'
+                                        )
                                     elif node.type == IndentationTypes.IMPLICIT:
                                         self.add_issue(part, 135, 'xxx')
                                     else:
-                                        self.add_issue(part, 121, 'Continuation line under-indented for hanging indent')
+                                        self.add_issue(
+                                            part,
+                                            121,
+                                            'Continuation line under-indented for hanging indent'
+                                        )
                                 else:
                                     if node.type == IndentationTypes.VERTICAL_BRACKET:
-                                        self.add_issue(part, 127, 'Continuation line over-indented for visual indent')
+                                        self.add_issue(
+                                            part,
+                                            127,
+                                            'Continuation line over-indented for visual indent'
+                                        )
                                     elif node.type == IndentationTypes.IMPLICIT:
                                         self.add_issue(part, 136, 'xxx')
                                     else:
-                                        self.add_issue(part, 126, 'Continuation line over-indented for hanging indent')
+                                        self.add_issue(
+                                            part,
+                                            126,
+                                            'Continuation line over-indented for hanging indent'
+                                        )
         else:
             self._check_spacing(part, spacing)
 
@@ -524,7 +553,7 @@ class PEP8Normalizer(ErrorFinder):
         else:
             last_column = part.end_pos[1]
         if last_column > self._config.max_characters \
-                and spacing.start_pos[1] <= self._config.max_characters :
+                and spacing.start_pos[1] <= self._config.max_characters:
             # Special case for long URLs in multi-line docstrings or comments,
             # but still report the error when the 72 first chars are whitespaces.
             report = True
@@ -538,7 +567,7 @@ class PEP8Normalizer(ErrorFinder):
                     part,
                     501,
                     'Line too long (%s > %s characters)' %
-                        (last_column, self._config.max_characters),
+                    (last_column, self._config.max_characters),
                 )
 
     def _check_spacing(self, part, spacing):
@@ -573,11 +602,11 @@ class PEP8Normalizer(ErrorFinder):
                 message = "Whitespace before '%s'" % part.value
                 add_if_spaces(spacing, 202, message)
             elif part in (',', ';') or part == ':' \
-                    and part.parent.type not in  _POSSIBLE_SLICE_PARENTS:
+                    and part.parent.type not in _POSSIBLE_SLICE_PARENTS:
                 message = "Whitespace before '%s'" % part.value
                 add_if_spaces(spacing, 203, message)
             elif prev == ':' and prev.parent.type in _POSSIBLE_SLICE_PARENTS:
-                pass # TODO
+                pass  # TODO
             elif prev in (',', ';', ':'):
                 add_not_spaces(spacing, 231, "missing whitespace after '%s'")
             elif part == ':':  # Is a subscript
@@ -602,9 +631,17 @@ class PEP8Normalizer(ErrorFinder):
                     if param.type == 'param' and param.annotation:
                         add_not_spaces(spacing, 252, 'Expected spaces around annotation equals')
                     else:
-                        add_if_spaces(spacing, 251, 'Unexpected spaces around keyword / parameter equals')
+                        add_if_spaces(
+                            spacing,
+                            251,
+                            'Unexpected spaces around keyword / parameter equals'
+                        )
                 elif part in _BITWISE_OPERATOR or prev in _BITWISE_OPERATOR:
-                    add_not_spaces(spacing, 227, 'Missing whitespace around bitwise or shift operator')
+                    add_not_spaces(
+                        spacing,
+                        227,
+                        'Missing whitespace around bitwise or shift operator'
+                    )
                 elif part == '%' or prev == '%':
                     add_not_spaces(spacing, 228, 'Missing whitespace around modulo operator')
                 else:
@@ -621,8 +658,7 @@ class PEP8Normalizer(ErrorFinder):
 
                 if spaces and part not in _ALLOW_SPACE and prev not in _ALLOW_SPACE:
                     message_225 = 'Missing whitespace between tokens'
-                    #print('xy', spacing)
-                    #self.add_issue(spacing, 225, message_225)
+                    # self.add_issue(spacing, 225, message_225)
                     # TODO why only brackets?
                     if part in _OPENING_BRACKETS:
                         message = "Whitespace before '%s'" % part.value
@@ -664,7 +700,8 @@ class PEP8Normalizer(ErrorFinder):
                         self.add_issue(leaf, 711, message)
                         break
                     elif node.value in ('True', 'False'):
-                        message = "comparison to False/True should be 'if cond is True:' or 'if cond:'"
+                        message = "comparison to False/True should be " \
+                                  "'if cond is True:' or 'if cond:'"
                         self.add_issue(leaf, 712, message)
                         break
         elif leaf.value in ('in', 'is'):
@@ -680,6 +717,7 @@ class PEP8Normalizer(ErrorFinder):
                 indentation = re.match(r'[ \t]*', line).group(0)
                 start_pos = leaf.line + i, len(indentation)
                 # TODO check multiline indentation.
+                start_pos
         elif typ == 'endmarker':
             if self._newline_count >= 2:
                 self.add_issue(leaf, 391, 'Blank line at end of file')
@@ -694,7 +732,7 @@ class PEP8Normalizer(ErrorFinder):
             return
         if code in (901, 903):
             # 901 and 903 are raised by the ErrorFinder.
-            super(PEP8Normalizer, self).add_issue(node, code, message)
+            super().add_issue(node, code, message)
         else:
             # Skip ErrorFinder here, because it has custom behavior.
             super(ErrorFinder, self).add_issue(node, code, message)
@@ -718,7 +756,7 @@ class PEP8NormalizerConfig(ErrorFinderConfig):
 
 
 # TODO this is not yet ready.
-#@PEP8Normalizer.register_rule(type='endmarker')
+# @PEP8Normalizer.register_rule(type='endmarker')
 class BlankLineAtEnd(Rule):
     code = 392
     message = 'Blank line at end of file'
