@@ -1,22 +1,23 @@
 import re
 from codecs import BOM_UTF8
+from typing import Tuple
 
 from parso.python.tokenize import group
 
 unicode_bom = BOM_UTF8.decode('utf-8')
 
 
-class PrefixPart(object):
+class PrefixPart:
     def __init__(self, leaf, typ, value, spacing='', start_pos=None):
         assert start_pos is not None
         self.parent = leaf
         self.type = typ
         self.value = value
         self.spacing = spacing
-        self.start_pos = start_pos
+        self.start_pos: Tuple[int, int] = start_pos
 
     @property
-    def end_pos(self):
+    def end_pos(self) -> Tuple[int, int]:
         if self.value.endswith('\n'):
             return self.start_pos[0] + 1, 0
         if self.value == unicode_bom:
@@ -71,7 +72,7 @@ def split_prefix(leaf, start_pos):
     value = spacing = ''
     bom = False
     while start != len(leaf.prefix):
-        match =_regex.match(leaf.prefix, start)
+        match = _regex.match(leaf.prefix, start)
         spacing = match.group(1)
         value = match.group(2)
         if not value:
