@@ -1,30 +1,32 @@
-from collections import namedtuple
 import re
 import sys
 from ast import literal_eval
 from functools import total_ordering
-
-from parso._compatibility import unicode
+from typing import NamedTuple, Sequence, Union
 
 # The following is a list in Python that are line breaks in str.splitlines, but
 # not in Python. In Python only \r (Carriage Return, 0xD) and \n (Line Feed,
 # 0xA) are allowed to split lines.
 _NON_LINE_BREAKS = (
-    u'\v',  # Vertical Tabulation 0xB
-    u'\f',  # Form Feed 0xC
-    u'\x1C',  # File Separator
-    u'\x1D',  # Group Separator
-    u'\x1E',  # Record Separator
-    u'\x85',  # Next Line (NEL - Equivalent to CR+LF.
-              # Used to mark end-of-line on some IBM mainframes.)
-    u'\u2028',  # Line Separator
-    u'\u2029',  # Paragraph Separator
+    '\v',  # Vertical Tabulation 0xB
+    '\f',  # Form Feed 0xC
+    '\x1C',  # File Separator
+    '\x1D',  # Group Separator
+    '\x1E',  # Record Separator
+    '\x85',  # Next Line (NEL - Equivalent to CR+LF.
+             # Used to mark end-of-line on some IBM mainframes.)
+    '\u2028',  # Line Separator
+    '\u2029',  # Paragraph Separator
 )
 
-Version = namedtuple('Version', 'major, minor, micro')
+
+class Version(NamedTuple):
+    major: int
+    minor: int
+    micro: int
 
 
-def split_lines(string, keepends=False):
+def split_lines(string: str, keepends: bool = False) -> Sequence[str]:
     r"""
     Intended for Python code. In contrast to Python's :py:meth:`str.splitlines`,
     looks at form feeds and other special characters as normal text. Just
@@ -68,7 +70,9 @@ def split_lines(string, keepends=False):
         return re.split(r'\n|\r\n|\r', string)
 
 
-def python_bytes_to_unicode(source, encoding='utf-8', errors='strict'):
+def python_bytes_to_unicode(
+    source: Union[str, bytes], encoding: str = 'utf-8', errors: str = 'strict'
+) -> str:
     """
     Checks for unicode BOMs and PEP 263 encoding declarations. Then returns a
     unicode object like in :py:meth:`bytes.decode`.
@@ -92,33 +96,33 @@ def python_bytes_to_unicode(source, encoding='utf-8', errors='strict'):
         possible_encoding = re.search(br"coding[=:]\s*([-\w.]+)",
                                       first_two_lines)
         if possible_encoding:
-            return possible_encoding.group(1)
+            e = possible_encoding.group(1)
+            if not isinstance(e, str):
+                e = str(e, 'ascii', 'replace')
+            return e
         else:
             # the default if nothing else has been set -> PEP 263
             return encoding
 
-    if isinstance(source, unicode):
+    if isinstance(source, str):
         # only cast str/bytes
         return source
 
     encoding = detect_encoding()
-    if not isinstance(encoding, unicode):
-        encoding = unicode(encoding, 'utf-8', 'replace')
-
     try:
         # Cast to unicode
-        return unicode(source, encoding, errors)
+        return str(source, encoding, errors)
     except LookupError:
         if errors == 'replace':
             # This is a weird case that can happen if the given encoding is not
             # a valid encoding. This usually shouldn't happen with provided
             # encodings, but can happen if somebody uses encoding declarations
             # like `# coding: foo-8`.
-            return unicode(source, 'utf-8', errors)
+            return str(source, 'utf-8', errors)
         raise
 
 
-def version_info():
+def version_info() -> Version:
     """
     Returns a namedtuple of parso's version, similar to Python's
     ``sys.version_info``.
@@ -128,8 +132,35 @@ def version_info():
     return Version(*[x if i == 3 else int(x) for i, x in enumerate(tupl)])
 
 
-def _parse_version(version):
-    match = re.match(r'(\d+)(?:\.(\d{1,2})(?:\.\d+)?)?$', version)
+class _PythonVersionInfo(NamedTuple):
+    major: int
+    minor: int
+
+
+@total_ordering
+class PythonVersionInfo(_PythonVersionInfo):
+    def __gt__(self, other):
+        if isinstance(other, tuple):
+            if len(other) != 2:
+                raise ValueError("Can only compare to tuples of length 2.")
+            return (self.major, self.minor) > other
+        super().__gt__(other)
+
+        return (self.major, self.minor)
+
+    def __eq__(self, other):
+        if isinstance(other, tuple):
+            if len(other) != 2:
+                raise ValueError("Can only compare to tuples of length 2.")
+            return (self.major, self.minor) == other
+        super().__eq__(other)
+
+    def __ne__(self, other):
+        return not self.__eq__(other)
+
+
+def _parse_version(version) -> PythonVersionInfo:
+    match = re.match(r'(\d+)(?:\.(\d{1,2})(?:\.\d+)?)?((a|b|rc)\d)?$', version)
     if match is None:
         raise ValueError('The given version is not in the right format. '
                          'Use something like "3.8" or "3".')
@@ -149,37 +180,15 @@ def _parse_version(version):
     return PythonVersionInfo(major, minor)
 
 
-@total_ordering
-class PythonVersionInfo(namedtuple('Version', 'major, minor')):
-    def __gt__(self, other):
-        if isinstance(other, tuple):
-            if len(other) != 2:
-                raise ValueError("Can only compare to tuples of length 2.")
-            return (self.major, self.minor) > other
-        super(PythonVersionInfo, self).__gt__(other)
-
-        return (self.major, self.minor)
-
-    def __eq__(self, other):
-        if isinstance(other, tuple):
-            if len(other) != 2:
-                raise ValueError("Can only compare to tuples of length 2.")
-            return (self.major, self.minor) == other
-        super(PythonVersionInfo, self).__eq__(other)
-
-    def __ne__(self, other):
-        return not self.__eq__(other)
-
-
-def parse_version_string(version=None):
+def parse_version_string(version: str = None) -> PythonVersionInfo:
     """
-    Checks for a valid version number (e.g. `3.8` or `2.7.1` or `3`) and
+    Checks for a valid version number (e.g. `3.8` or `3.10.1` or `3`) and
     returns a corresponding version info that is always two characters long in
     decimal.
     """
     if version is None:
         version = '%s.%s' % sys.version_info[:2]
-    if not isinstance(version, (unicode, str)):
+    if not isinstance(version, str):
         raise TypeError('version must be a string like "3.8"')
 
     return _parse_version(version)
