@@ -13,7 +13,7 @@ Parso consists of a small API to parse Python and analyse the syntax tree.
 A simple example:
 
 >>> import parso
->>> module = parso.parse('hello + 1', version="3.6")
+>>> module = parso.parse('hello + 1', version="3.9")
 >>> expr = module.children[0]
 >>> expr
 PythonNode(arith_expr, [<Name: hello@1,0>, <Operator: +>, <Number: 1>])
@@ -43,7 +43,7 @@ from parso.grammar import Grammar, load_grammar
 from parso.utils import split_lines, python_bytes_to_unicode
 
 
-__version__ = '0.7.0'
+__version__ = '0.8.1'
 
 
 def parse(code=None, **kwargs):
