@@ -3,7 +3,6 @@ from textwrap import dedent
 
 import pytest
 
-from parso._compatibility import u
 from parso import parse
 from parso.python import tree
 from parso.utils import split_lines
@@ -110,23 +109,15 @@ def test_param_splitting(each_version):
     but Jedi does this to simplify argument parsing.
     """
     def check(src, result):
-        # Python 2 tuple params should be ignored for now.
         m = parse(src, version=each_version)
-        if each_version.startswith('2'):
-            # We don't want b and c to be a part of the param enumeration. Just
-            # ignore them, because it's not what we want to support in the
-            # future.
-            func = next(m.iter_funcdefs())
-            assert [param.name.value for param in func.get_params()] == result
-        else:
-            assert not list(m.iter_funcdefs())
+        assert not list(m.iter_funcdefs())
 
     check('def x(a, (b, c)):\n pass', ['a'])
     check('def x((b, c)):\n pass', [])
 
 
 def test_unicode_string():
-    s = tree.String(None, u('bö'), (0, 0))
+    s = tree.String(None, 'bö', (0, 0))
     assert repr(s)  # Should not raise an Error!
 
 
@@ -135,17 +126,8 @@ def test_backslash_dos_style(each_version):
 
 
 def test_started_lambda_stmt(each_version):
-    m = parse(u'lambda a, b: a i', version=each_version)
+    m = parse('lambda a, b: a i', version=each_version)
     assert m.children[0].type == 'error_node'
-
-
-def test_python2_octal(each_version):
-    module = parse('0660', version=each_version)
-    first = module.children[0]
-    if each_version.startswith('2'):
-        assert first.type == 'number'
-    else:
-        assert first.type == 'error_node'
 
 
 @pytest.mark.parametrize('code', ['foo "', 'foo """\n', 'foo """\nbar'])
@@ -195,6 +177,11 @@ def test_named_expression(works_ge_py38):
     works_ge_py38.parse("(a := 1, a + 1)")
 
 
+def test_extended_rhs_annassign(works_ge_py38):
+    works_ge_py38.parse("x: y = z,")
+    works_ge_py38.parse("x: Tuple[int, ...] = z, *q, w")
+
+
 @pytest.mark.parametrize(
     'param_code', [
         'a=1, /',
@@ -208,6 +195,7 @@ def test_named_expression(works_ge_py38):
 )
 def test_positional_only_arguments(works_ge_py38, param_code):
     works_ge_py38.parse("def x(%s): pass" % param_code)
+
 
 @pytest.mark.parametrize(
     'expression', [
