@@ -1,6 +1,5 @@
 # -*- coding: utf-8    # This file contains Unicode characters.
 
-import sys
 from textwrap import dedent
 
 import pytest
@@ -31,7 +30,7 @@ FSTRING_END = PythonTokenTypes.FSTRING_END
 def _get_token_list(string, version=None):
     # Load the current version.
     version_info = parse_version_string(version)
-    return list(tokenize.tokenize(string, version_info))
+    return list(tokenize.tokenize(string, version_info=version_info))
 
 
 def test_end_pos_one_line():
@@ -108,7 +107,7 @@ def test_tokenize_multiline_I():
     fundef = '''""""\n'''
     token_list = _get_token_list(fundef)
     assert token_list == [PythonToken(ERRORTOKEN, '""""\n', (1, 0), ''),
-                          PythonToken(ENDMARKER ,       '', (2, 0), '')]
+                          PythonToken(ENDMARKER, '', (2, 0), '')]
 
 
 def test_tokenize_multiline_II():
@@ -117,7 +116,7 @@ def test_tokenize_multiline_II():
     fundef = '''""""'''
     token_list = _get_token_list(fundef)
     assert token_list == [PythonToken(ERRORTOKEN, '""""', (1, 0), ''),
-                          PythonToken(ENDMARKER,      '', (1, 4), '')]
+                          PythonToken(ENDMARKER, '', (1, 4), '')]
 
 
 def test_tokenize_multiline_III():
@@ -126,7 +125,7 @@ def test_tokenize_multiline_III():
     fundef = '''""""\n\n'''
     token_list = _get_token_list(fundef)
     assert token_list == [PythonToken(ERRORTOKEN, '""""\n\n', (1, 0), ''),
-                          PythonToken(ENDMARKER,          '', (3, 0), '')]
+                          PythonToken(ENDMARKER, '', (3, 0), '')]
 
 
 def test_identifier_contains_unicode():
@@ -136,12 +135,7 @@ def test_identifier_contains_unicode():
     ''')
     token_list = _get_token_list(fundef)
     unicode_token = token_list[1]
-    if sys.version_info.major >= 3:
-        assert unicode_token[0] == NAME
-    else:
-        # Unicode tokens in Python 2 seem to be identified as operators.
-        # They will be ignored in the parser, that's ok.
-        assert unicode_token[0] == ERRORTOKEN
+    assert unicode_token[0] == NAME
 
 
 def test_quoted_strings():
@@ -184,19 +178,16 @@ def test_ur_literals():
             assert typ == NAME
 
     check('u""')
-    check('ur""', is_literal=not sys.version_info.major >= 3)
-    check('Ur""', is_literal=not sys.version_info.major >= 3)
-    check('UR""', is_literal=not sys.version_info.major >= 3)
+    check('ur""', is_literal=False)
+    check('Ur""', is_literal=False)
+    check('UR""', is_literal=False)
     check('bR""')
-    # Starting with Python 3.3 this ordering is also possible.
-    if sys.version_info.major >= 3:
-        check('Rb""')
+    check('Rb""')
 
-    # Starting with Python 3.6 format strings where introduced.
-    check('fr""', is_literal=sys.version_info >= (3, 6))
-    check('rF""', is_literal=sys.version_info >= (3, 6))
-    check('f""', is_literal=sys.version_info >= (3, 6))
-    check('F""', is_literal=sys.version_info >= (3, 6))
+    check('fr""')
+    check('rF""')
+    check('f""')
+    check('F""')
 
 
 def test_error_literal():
@@ -229,9 +220,6 @@ def test_endmarker_end_pos():
     check('a\\')
 
 
-xfail_py2 = dict(marks=[pytest.mark.xfail(sys.version_info[0] == 2, reason='Python 2')])
-
-
 @pytest.mark.parametrize(
     ('code', 'types'), [
         # Indentation
@@ -243,12 +231,10 @@ xfail_py2 = dict(marks=[pytest.mark.xfail(sys.version_info[0] == 2, reason='Pyth
 
         # Name stuff
         ('1foo1', [NUMBER, NAME]),
-        pytest.param(
-            u'மெல்லினம்', [NAME],
-            **xfail_py2),
-        pytest.param(u'²', [ERRORTOKEN], **xfail_py2),
-        pytest.param(u'ä²ö', [NAME, ERRORTOKEN, NAME], **xfail_py2),
-        pytest.param(u'ää²¹öö', [NAME, ERRORTOKEN, NAME], **xfail_py2),
+        ('மெல்லினம்', [NAME]),
+        ('²', [ERRORTOKEN]),
+        ('ä²ö', [NAME, ERRORTOKEN, NAME]),
+        ('ää²¹öö', [NAME, ERRORTOKEN, NAME]),
         (' \x00a', [INDENT, ERRORTOKEN, NAME, DEDENT]),
         (dedent('''\
             class BaseCache:
@@ -411,8 +397,8 @@ def test_backslash():
         ]),
     ]
 )
-def test_fstring_token_types(code, types, version_ge_py36):
-    actual_types = [t.type for t in _get_token_list(code, version_ge_py36)]
+def test_fstring_token_types(code, types, each_version):
+    actual_types = [t.type for t in _get_token_list(code, each_version)]
     assert types + [ENDMARKER] == actual_types
 
 
